feat: settings + refactor

+++ apps/api/app/controller/profile.controller.ts
@@ -12,19 +12,19 @@ import NotFoundError from '../helpers/errors/not-found.error';
 export default {
   async getOne(req: Request, res: Response) {
     const [profileId] = checkParams(req.params.profileId);
-    const profile = await Profile.findOne(profileId);
+    const profile = await Profile.findOne({ id: profileId });
 
     return res.status(200).json(profile);
   },
   async createOne(req: Request, res: Response) {
     const data = req.body;
-    const result = await Profile.create(data);
+    const result = await Profile.createOne(data);
 
     return res.status(201).json({ success: result });
   },
   async updateOne(req: Request, res: Response) {
-    const data = req.body.data;
-    const isUpdate = await Profile.updateProfile(data);
+    const { profile_id, ...data } = req.body.data;
+    const isUpdate = await Profile.updateOne({ id: profile_id }, data);
 
     return res.status(204).send({ success: isUpdate });
   },
@@ -46,8 +46,12 @@ export default {
     // return new image url
 
     if (!avatarImage) throw new UserInputError('No image provided');
+    const profile = await Profile.findOne({ id: Number(profile_id) });
 
-    const { username, avatar_url } = await Profile.findOne(Number(profile_id));
+    if (!profile) {
+      throw new UserInputError('Profile not found');
+    }
+    const { username, avatar_url } = profile;
 
     avatarImage.originalname = `avatar_${username}`;
 
@@ -55,18 +59,23 @@ export default {
       height: WIDTH_AVATAR,
       width: WIDTH_AVATAR,
     });
-    await Image.create({
+    await Image.createOne({
       url: link,
       key: key,
       size: WIDTH_AVATAR,
     });
 
     if (avatar_url) {
-      const [imageToDelete] = await Image.findBy({ url: avatar_url });
-      await Image.delete(imageToDelete.id);
-      await deleteImageFromBucket(imageToDelete.key);
+      const imageToDelete = await Image.findOne({ url: avatar_url });
+      if (!imageToDelete) throw new NotFoundError('Image not found');
+      await Image.deleteOne({ id: imageToDelete.id });
+
+      if (imageToDelete.key) {
+        await deleteImageFromBucket(imageToDelete.key);
+      }
     }
-    await Profile.updateProfile({ profile_id, avatar_url: link });
+
+    await Profile.updateOne({ id: profile_id }, { avatar_url: link });
 
     return res.status(200).json({ link });
   },
